M5T12 split utils into several files
