--- src/utils.js
@@ -1,57 +1,0 @@
-import {RENDER_POSITION} from "./const";
-
-const getCurrentDate = () => {
-  const currentDate = new Date();
-  currentDate.setHours(23, 59, 59, 999);
-
-  return currentDate;
-};
-
-export const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const isExpired = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
-
-  return getCurrentDate().getTime() > dueDate.getTime();
-};
-
-export const isExpiringToday = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
-
-  return getCurrentDate().getTime() === dueDate.getTime();
-};
-
-export const isRepeating = (repeatingDays) => {
-  return Object.values(repeatingDays).some(Boolean);
-};
-
-export const getFormattingDueDate = (date) => {
-  return date.toLocaleString(`ru-RU`, {day: `numeric`, month: `long`});
-};
-
-export const renderElement = (container, element, place) => {
-  switch (place) {
-    case RENDER_POSITION.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RENDER_POSITION.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
